Divided functions save, menu,, game_logic
